requirements.txt
sphinx
httplib2
python-dateutil
